+++ paper/paper.bib
@@ -14,7 +14,8 @@
   booktitle = {Proceedings of the 2020 Conference on Empirical Methods in Natural Language Processing (EMNLP)},
   pages = {671--683},
   year = {2020},
-  organization = {Association for Computational Linguistics}
+  organization = {Association for Computational Linguistics},
+  doi = "10.18653/v1/2020.emnlp-main.49"
 }
 
 @article{datta2024consistency,
@@ -31,7 +32,8 @@
   journal = {Proceedings of the 61st Annual Meeting of the Association for Computational Linguistics (Volume 1: Long Papers)},
   pages = {14014--14031},
   year = {2023},
-  organization = {Association for Computational Linguistics}
+  organization = {Association for Computational Linguistics},
+  doi = "10.18653/v1/2023.acl-long.783"
 }
 
 @article{freivalds2002learning,
@@ -47,7 +49,8 @@
   title={POE: Process of Elimination for Multiple Choice Reasoning},
   author={Ma, Chenkai and Du, Xinya},
   journal={arXiv preprint arXiv:2310.15575},
-  year={2023}
+  year={2023},
+  doi = "10.18653/v1/2023.emnlp-main.273"
 }
 
 @inproceedings{holtzman2021surface,
@@ -56,7 +59,8 @@
   booktitle = {Proceedings of the 2021 Conference on Empirical Methods in Natural Language Processing},
   pages = {7038--7051},
   year = {2021},
-  organization = {Association for Computational Linguistics}
+  organization = {Association for Computational Linguistics},
+  doi = "10.18653/v1/2021.emnlp-main.564"
 }
 
 @inproceedings{kojima2022large,
@@ -78,7 +82,8 @@
   title = {Z-ICL: Zero-shot in-context learning with pseudo-demonstrations},
   author = {Lyu, Xinxi and Min, Sewon and Beltagy, Iz and Zettlemoyer, Luke and Hajishirzi, Hannaneh},
   booktitle = {ICLR 2023 Workshop on Mathematical and Empirical Understanding of Foundation Models},
-  year = {2023}
+  year = {2023},
+  doi = "10.18653/v1/2023.acl-long.129"
 }
 
 @article{malkin2022coherence,
@@ -87,7 +92,8 @@
   journal = {Proceedings of the 60th Annual Meeting of the Association for Computational Linguistics (Volume 1: Long Papers)},
   pages = {8214--8236},
   year = {2022},
-  organization = {Association for Computational Linguistics}
+  organization = {Association for Computational Linguistics},
+  doi = "10.18653/v1/2022.acl-long.565"
 }
 
 @inproceedings{min2022noisy,
@@ -96,7 +102,8 @@
   booktitle = {Proceedings of the 60th Annual Meeting of the Association for Computational Linguistics (Volume 1: Long Papers)},
   pages = {5316--5330},
   year = {2022},
-  organization = {Association for Computational Linguistics}
+  organization = {Association for Computational Linguistics},
+  doi = "10.18653/v1/2022.acl-long.365"
 }
 
 @inproceedings{nie2020adversarial,
@@ -105,7 +112,8 @@
   booktitle = {Proceedings of the 58th Annual Meeting of the Association for Computational Linguistics},
   pages = {4885--4901},
   year = {2020},
-  organization = {Association for Computational Linguistics}
+  organization = {Association for Computational Linguistics},
+  doi = "10.18653/v1/2020.acl-main.441"
 }
 
 @inproceedings{ouyang2022training,
@@ -128,7 +136,8 @@
[…]
