### 2024-11-21 #7391 @faroit
@hellokayas 👋 i read the paper and tried to find suitable reviewers based on the content in our JOSS reviewer database. I have a few potential reviewers in mind:

- @ hoechenberger
- @ stain
- @ jorgedch
- @ Mind-the-Cap
- @ mitchellg

@hellokayas, I am also interested in your input on potential reviewers, so any suggestions you have are also welcome. For this, please check the JOSS [COI policy](<https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy>) and only mention them with a space between the @ and their handle (so you would refer to me as "@ faroit").
