### 2024-11-21 #7391 @faroit
@hellokayas in the meantime, I suggest that you address a few of the missing DOIs in the references, if not already something you worked on
