### 2024-11-23 #7391 @hellokayas
> @hellokayas in the meantime, I suggest that you address a few of the missing DOIs in the references, if not already something you worked on

@faroit We have already worked on fixing the missing DOIs. Please check.
