+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume = {33},
   pages = {1877--1901},
   year = {2020},
-  organization = {Curran Associates, Inc.}
+  organization = {Curran Associates, Inc.},
+  doi = "10.48550/arXiv.2005.14165"
 }
 
 @inproceedings{du2020event,
@@ -23,7 +24,8 @@
   author={Datta, Arghya and Chakrabarty, Sayak},
   journal={Advances in Neural Information Processing Systems},
   volume={36},
-  year={2024}
+  year={2024},
+  doi = "10.48550/arXiv.2311.05046"
 }
 
 @article{fei2023mitigating,
@@ -42,7 +44,8 @@
   journal = {Inf. Comput.},
   volume = {176},
   pages = {37--50},
-  year = {2002}
+  year = {2002},
+  doi = "10.1006/inco.2001.2922"
 }
 
 @article{ma2023poe,
@@ -67,15 +70,17 @@
   title = {Large language models are zero-shot reasoners},
   author = {Kojima, Takeshi and Gu, Shixiang Shane and Reid, Machel and Matsuo, Yutaka and Iwasawa, Yusuke},
   booktitle = {Advances in Neural Information Processing Systems},
-  year = {2022}
+  year = {2022},
+  doi = "10.48550/arXiv.2205.11916"
 }
 
 @article{zhang2023dynamically,
-  title={: A Dynamically Adaptive Defense to a Novel Attack on Review Fraud Detection Engines},
+  title={SockDef: A Dynamically Adaptive Defense to a Novel Attack on Review Fraud Detection Engines},
   author={Zhang, Youzhi and Chakrabarty, Sayak and Liu, Rui and Pugliese, Andrea and Subrahmanian, VS},
   journal={IEEE Transactions on Computational Social Systems},
   year={2023},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = "10.1109/TCSS.2023.3321345"
 }
 
 @inproceedings{lyu2023z,
@@ -120,14 +125,16 @@
   title = {Training language models to follow instructions with human feedback},
   author = {Ouyang, Long and Wu, Jeffrey and Jiang, Xu and Almeida, Diogo and Wainwright, Carroll and Mishkin, Pamela and Zhang, Chong and Agarwal, Sandhini and Slama, Katarina and Gray, Alex and others},
   booktitle = {Advances in Neural Information Processing Systems},
-  year = {2022}
+  year = {2022},
+  doi = "10.48550/arXiv.2203.02155"
 }
 
 @inproceedings{robinson2023leveraging,
   title = {Leveraging large language models for multiple choice question answering},
   author = {Robinson, Joshua and Wingate, David},
   booktitle = {The Eleventh International Conference on Learning Representations},
-  year = {2023}
+  year = {2023},
+  doi = "10.48550/arXiv.2210.12353"
 }
 
 @inproceedings{sap2019social,
@@ -162,7 +169,8 @@
   title = {Beyond the imitation game: Quantifying and extrapolating the capabilities of language models},
   author = {Srivastava, Aarohi and Rastogi, Abhinav and Rao, Abhishek and Shoeb, Abu Awal Md and Abid, Abubakar and Fisch, Adam and Brown, Adam R and Santoro, Adam and Gupta, Aditya and Garriga-Alonso, Adrià and others},
   booktitle = {Transactions on Machine Learning Research},
-  year = {2023}
+  year = {2023},
+  doi = "10.48550/arXiv.2206.04615"
 }
 
 @inproceedings{suzgun2023challenging,
@@ -181,49 +189,56 @@
[…]
