### 2024-12-07 #7391 @hellokayas
> > We have already worked on fixing the missing DOIs. Please check
>
> @hellokayas it looks like there are still a couple of references that where the DOI is missing and that probably have one. E.g most acl papers should be on https://aclanthology.org and for some others I found at least an arxiv version that could be references alternatively.

@faroit We have worked on the couple of missing DOIs. Can you check and let us know if everything looks good now?
