### 2025-01-26 #7391 @jorgedch
## Pre-Review checklist for @jorgedch

To be fair, @souradipp76 has submitted around 4/5 of the code to the project repo (@souradipp76: 21581-12908=8673, @hellokayas: 2730-314=2416). Of course, amount of code not always reflects the actual contribution to the project, but it would be nice if @hellokayas could elaborate why @souradipp76 is not the submitting author.

1. The documentation should include some rough estimates of the following:
- size of the data downloaded by `data_downloaders.sh`
- size of the data downloaded by `model_downloaders.sh`
- a full list of the tests downloaded
    - what each test is about
    - what results to expect from each test
    - time required to run each test
2. `data_downloaders.sh` works correctly
3. `model_downloaders.sh` fails with the following error:

```
  File "/content/MM-PoE/mm_poe/models/model_downloaders/model_downloaders.py", line 204, in <module>
  File "/content/MM-PoE/mm_poe/models/model_downloaders/model_downloaders.py", line 179, in main
    tokenizer = tokenizer_func.from_pretrained(checkpoint)
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/models/auto/processing_auto.py", line 316, in from_pretrained
    return processor_class.from_pretrained(
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/processing_utils.py", line 465, in from_pretrained
    args = cls._get_arguments_from_pretrained(pretrained_model_name_or_path, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/processing_utils.py", line 511, in _get_arguments_from_pretrained
    args.append(attribute_class.from_pretrained(pretrained_model_name_or_path, **kwargs))
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/models/auto/tokenization_auto.py", line 880, in from_pretrained
    return tokenizer_class.from_pretrained(pretrained_model_name_or_path, *inputs, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/tokenization_utils_base.py", line 2110, in from_pretrained
    return cls._from_pretrained(
           ^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/tokenization_utils_base.py", line 2336, in _from_pretrained
    tokenizer = cls(*init_inputs, **init_kwargs)
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/models/gpt2/tokenization_gpt2_fast.py", line 100, in __init__
    super().__init__(
  File "/usr/local/lib/python3.11/dist-packages/transformers/tokenization_utils_fast.py", line 114, in __init__
    fast_tokenizer = TokenizerFast.from_file(fast_tokenizer_file)
                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
Exception: data did not match any variant of untagged enum ModelWrapper at line 250373 column 3
```

Will be able to assess once the tests are running correctly.

Will be able to assess once the tests are running correctly.

The installation instructions work correctly. They just should mention explicitly that the dependencies require CUDA-compatible hardware to run the project.

The example tests fail in the following places:
1. The docs instructo to run `$ bash 7_main_exp.sh`, but the script name is different from the downloaded one (`7_main_exp_vqa.sh`) and the test fails.
2. `7_main_exp_vqa.sh` produces the following error:

```
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 399, in cached_file
    resolved_file = hf_hub_download(
                    ^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/huggingface_hub/utils/_validators.py", line 106, in _inner_fn
    validate_repo_id(arg_value)
  File "/usr/local/lib/python3.11/dist-packages/huggingface_hub/utils/_validators.py", line 154, in validate_repo_id
    raise HFValidationError(
huggingface_hub.errors.HFValidationError: Repo id must be in the form 'repo_name' or 'namespace/repo_name': '/content/models/BLIP2/Salesforce/blip2-opt-2.7b'. Use `repo_type` argument if needed.

  File "/content/MM-PoE/mm_poe/methods/vision_language_modeling.py", line 370, in <module>
  File "/content/MM-PoE/mm_poe/methods/vision_language_modeling.py", line 81, in main
    model, tokenizer = load_model(device, model_path, args)
                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/content/MM-PoE/mm_poe/methods/utils/utils.py", line 591, in load_model
    tokenizer = tokenizer_func.from_pretrained(model_path)
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/models/auto/processing_auto.py", line 242, in from_pretrained
    processor_config_file = get_file_from_repo(
                            ^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 551, in get_file_from_repo
    return cached_file(
           ^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 463, in cached_file
    raise EnvironmentError(
OSError: Incorrect path_or_model_id: '/content/models/BLIP2/Salesforce/blip2-opt-2.7b'. Please provide either the path to a local folder or the repo_id of a model on the Hub.
```

3. `bash 9_mask_vqa.sh` produces the following error:

```
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 399, in cached_file
    resolved_file = hf_hub_download(
                    ^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/huggingface_hub/utils/_validators.py", line 106, in _inner_fn
    validate_repo_id(arg_value)
  File "/usr/local/lib/python3.11/dist-packages/huggingface_hub/utils/_validators.py", line 154, in validate_repo_id
    raise HFValidationError(
huggingface_hub.errors.HFValidationError: Repo id must be in the form 'repo_name' or 'namespace/repo_name': '/content/models/BLIP2/Salesforce/blip2-opt-2.7b'. Use `repo_type` argument if needed.

  File "/content/MM-PoE/mm_poe/methods/process_of_elimination_vqa.py", line 418, in <module>
  File "/content/MM-PoE/mm_poe/methods/process_of_elimination_vqa.py", line 72, in main
    model, tokenizer = load_model(device, model_path, args)
                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/content/MM-PoE/mm_poe/methods/utils/utils.py", line 591, in load_model
    tokenizer = tokenizer_func.from_pretrained(model_path)
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/models/auto/processing_auto.py", line 242, in from_pretrained
    processor_config_file = get_file_from_repo(
                            ^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 551, in get_file_from_repo
    return cached_file(
           ^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 463, in cached_file
    raise EnvironmentError(
OSError: Incorrect path_or_model_id: '/content/models/BLIP2/Salesforce/blip2-opt-2.7b'. Please provide either the path to a local folder or the repo_id of a model on the Hub.
```

4. `bash 11_few_shot_vqa.sh` produces the following error:

```
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 399, in cached_file
    resolved_file = hf_hub_download(
                    ^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/huggingface_hub/utils/_validators.py", line 106, in _inner_fn
    validate_repo_id(arg_value)
  File "/usr/local/lib/python3.11/dist-packages/huggingface_hub/utils/_validators.py", line 154, in validate_repo_id
    raise HFValidationError(
huggingface_hub.errors.HFValidationError: Repo id must be in the form 'repo_name' or 'namespace/repo_name': '/content/models/BLIP2/Salesforce/blip2-opt-2.7b'. Use `repo_type` argument if needed.

  File "/content/MM-PoE/mm_poe/methods/vision_language_modeling.py", line 370, in <module>
  File "/content/MM-PoE/mm_poe/methods/vision_language_modeling.py", line 81, in main
    model, tokenizer = load_model(device, model_path, args)
                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/content/MM-PoE/mm_poe/methods/utils/utils.py", line 591, in load_model
    tokenizer = tokenizer_func.from_pretrained(model_path)
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/models/auto/processing_auto.py", line 242, in from_pretrained
    processor_config_file = get_file_from_repo(
                            ^^^^^^^^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 551, in get_file_from_repo
    return cached_file(
           ^^^^^^^^^^^^
  File "/usr/local/lib/python3.11/dist-packages/transformers/utils/hub.py", line 463, in cached_file
    raise EnvironmentError(
OSError: Incorrect path_or_model_id: '/content/models/BLIP2/Salesforce/blip2-opt-2.7b'. Please provide either the path to a local folder or the repo_id of a model on the Hub.
```

5. There are more downloaded tests, but I only tried running the ones mentioned in the README file. All should be checked to run correctly when the ones above are fixed.

6.  It is not trivial to release software that people can test in arbitrary hardware, so an alternative to make things easier could be to include a Colab notebook, an AWS public AMI, or something similar. Then it is guaranteed that the reviewers will get the same results as the authors.

The documentation only shows how to run existing test scripts, but there is no information about:
- format requirements to use own dataset
- once a new dataset is ready, how to process it with the package

Basically, the current project provides a high-level testing solution that can be run via shell scripts, but there is no documentation on how to use the code as an actual python library by the community.

There are several tests in `MM-PoE/tests/test_cli.py`, but the tests are not documented, so it's not clear what they cover from the existing functionality. Some docstrings would help a lot.

Documentation about `2) Report issues or problems with the software`, and `3) Seek support` is missing.

`paper.md` contains all the required sections and some references, but the paper itself is still not compiled and `paper.bib` contains many more reference entries than the ones in the draft. Will be able to assess once the compiled PDF is added.
