+++ mm_poe/models/model_downloaders/model_downloaders.py
@@ -198,7 +198,8 @@ def main():
         for folder in folders_to_delete:
             print(f"Removing cached files at {folder}...")
             shutil.rmtree(folder)
-    print("Done")    
+    print("Done")
+
 
 if __name__ == "__main__":
     main()
