+++ .github/workflows/main.yml
@@ -86,7 +86,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install Pip
-        run: pip install --user --upgrade pip
+        run: python.exe -m pip install --user --upgrade pip
       - name: Install project
         run: pip install -e .[test]
       - name: run tests
