### 2025-01-31 36f0605
Added colab notebook for validation and updated README.
