+++ README.md
@@ -17,7 +17,11 @@ In the first step, the method evaluates and scores each option, systematically e
 
 Using this tool, researchers and practitioners can experiment and significantly improve the accuracy and reliability of VLMs in multiple choice reasoning tasks, making it a valuable tool for advancing machine learning models for visual reasoning.
 
-## Installing MM-PoE
+
+
+## Installation
+
+MM-PoE is available only on Linux/Windows. CUDA-compatible hardware is required to run the tool.
 
 ### Install it from PyPI
 
@@ -98,7 +102,81 @@ $ bash 9_mask_vqa.sh
 $ bash 11_few_shot_vqa.sh
 ```
 
-The numbers 7, 9 and 11 in the script names correspnds to the experiments related to MM-PoE using VLMs for multiple-choice visual question answering tasks and rest corresponds to experiments related to PoE using LLMs on logical resoning tasks. The results will be saved in `results/`.
+The numbers 7, 9 and 11 in the script names correspnds to the experiments related to MM-PoE using VLMs for multiple-choice visual question answering tasks and rest corresponds to experiments related to PoE using LLMs on logical resoning tasks. The results will be saved in `results/`. 
+
+Alternatively, run the notebook `scripts/experiments.ipynb` on Google Colab: <a src="https://colab.research.google.com/assets/colab-badge.svg" href="https://colab.research.google.com/github/souradipp76/MM-PoE/blob/main/scripts/experiments.ipynb" target="_blank" rel="noopener noreferrer"><img src="https://colab.research.google.com/assets/colab-badge.svg" alt="Open in Colab"></a>
+
+MM-PoE is compared against the following five baseline scoring methods to assess its relative performance:
+
+- **Language Modeling (LM)**: This baseline uses the raw vision language modeling likelihood as the scoring function.
+- **Average Language Modeling (AVG)**: This method averages the log probabilities across all tokens in the option.
+- **Calibration**: This involves adjusting the VLM scores based on calibration techniques that aim to correct for the model's confidence.
+- **Channel**: Channel methods score each option based on how likely the question is given the option, which reverses the typical conditional probability used in VLMs.
+- **Multiple Choice Prompting (MCP)**: This approach formats the input by presenting the question followed by all options, prompting the model to select the most likely option.
+
+| **Method**                | **Score**                                      | **Input**                                           | **Output** |
+|---------------------------|-----------------------------------------------|-----------------------------------------------------|------------|
+| LM                        | $\log P (y_i\|x,h)$                           | $x, \langle h \rangle$ the answer is:              | $y_1$      |
+| AVG                       | $1/len(y_i)*\log P(y_i\|x,h)$                  | $x, \langle h \rangle$ the answer is:              | $y_1$      |
+| Calibration               | $\log P(y_i\|x,h)$                            | $x, \langle h \rangle$ the answer is:              | $y_1$      |
+| Channel                   | $\log P(x\|y_i,h)$                            | $y_1, \langle h \rangle$                           | $x$ the answer is: |
[…]
