### 2025-02-01 8476a1e
Updated README and script.
