+++ mm_poe/models/model_downloaders/model_downloaders.py
@@ -1,4 +1,4 @@
-""" Model Downloaders """
+"""Model Downloaders"""
 
 import argparse
 import glob
