### 2025-02-01 f5c82b0
Added custom dataset format and image token for models.
