### 2025-02-02 1ce4a06
Updated the default model to GIT and seprated BLIp2 model families.
