+++ mm_poe/methods/vision_language_modeling.py
@@ -86,7 +86,7 @@ def main():
         preprocess_func = preprocess_function_seq2seq
         preprocess_func_channel = preprocess_function_seq2seq_channel
     elif args.model_family in [
-        "BLIP2",
+        "BLIP2-T5",
         "InstructBLIP",
         "PaliGemma",
         "Idefics2",
@@ -96,7 +96,7 @@ def main():
         preprocess_func_channel = preprocess_function_seq2seq_vqa_channel
         processor = tokenizer
         tokenizer = processor.tokenizer
-    elif args.model_family in ["GIT"]:
+    elif args.model_family in ["BLIP2-OPT", "GIT"]:
         compute_func = compute_conditional_score_causal_vqa
         preprocess_func = preprocess_function_causal_vqa
         preprocess_func_channel = preprocess_function_causal_vqa_channel
@@ -135,7 +135,8 @@ def main():
             "tokenizer": tokenizer,
         }
         if args.model_family in [
-            "BLIP2",
+            "BLIP2-OPT",
+            "BLIP2-T5",
             "InstructBLIP",
             "GIT",
             "PaliGemma",
@@ -227,7 +228,8 @@ def main():
                 "tokenizer": tokenizer,
             }
             if args.model_family in [
-                "BLIP2",
+                "BLIP2-OPT",
+                "BLIP2-T5",
                 "InstructBLIP",
                 "GIT",
                 "PaliGemma",
@@ -314,7 +316,8 @@ def main():
                 "tokenizer": tokenizer,
             }
             if args.model_family in [
-                "BLIP2",
+                "BLIP2-OPT",
+                "BLIP2-T5",
                 "InstructBLIP",
                 "GIT",
                 "PaliGemma",
