+++ tests/methods/utils/test_utils.py
@@ -448,7 +448,8 @@ def test_load_data_invalid_dataset():
         ("Dolly", "AutoModelForCausalLM", "AutoTokenizer"),
         ("T5", "AutoModelForSeq2SeqLM", "AutoTokenizer"),
         ("FLAN-T5", "AutoModelForSeq2SeqLM", "AutoTokenizer"),
-        ("BLIP2", "AutoModelForVision2Seq", "AutoProcessor"),
+        ("BLIP2-OPT", "AutoModelForVision2Seq", "AutoProcessor"),
+        ("BLIP2-T5", "AutoModelForVision2Seq", "AutoProcessor"),
         ("InstructBLIP", "AutoModelForVision2Seq", "AutoProcessor"),
         ("GIT", "AutoModelForVision2Seq", "AutoProcessor"),
         ("PaliGemma", "AutoModelForVision2Seq", "AutoProcessor"),
