### 2025-02-02 89ad982
Fixed device during model loading and inference calibration outputs.
