+++ mm_poe/methods/utils/utils.py
@@ -645,7 +645,6 @@ def load_model(device, model_path, args):
         )
     else:  # FP32
         model = model_func.from_pretrained(model_path, device_map=device)
-    model.to(device)
     print(
         f"Memory footprint: {model.get_memory_footprint() / 1024 **3:.2f} GB."
     )
