### 2025-02-02 #7391 @souradipp76
@jorgedch The issues have been addressed. Kindly verify.

The documentation should include some rough estimates of the following:
- size of the data downloaded by data_downloaders.sh -> **Links have been provided in README to each dataset where size information is available**
- size of the data downloaded by model_downloaders.sh -> **Links have been provided in README to each model where size information is available**
- a full list of the tests downloaded -> **Test information has been provided in [Running Experiments](https://github.com/souradipp76/MM-PoE#running-experiments) section in README**
  - what each test is about -> **Included in [Running Experiments](https://github.com/souradipp76/MM-PoE#running-experiments) section in README**
  - what results to expect from each test -> **Included in [Running Experiments](https://github.com/souradipp76/MM-PoE#running-experiments) section in README**
  - time required to run each test -> **Time to run each test varies based on the model, sample size, and batch size used. These are configurable parameters in the scripts.**

The documentation only shows how to run existing test scripts, but there is no information about:
- format requirements to use own dataset -> **Included in the [Supported Datasets](https://github.com/souradipp76/MM-PoE#supported-datasets) section in README**
- once a new dataset is ready, how to process it with the package -> **Included in the [Supported Datasets](https://github.com/souradipp76/MM-PoE#supported-datasets) section in README**

There are several tests in MM-PoE/tests/test_cli.py, but the tests are not documented, so it's not clear what they cover from the existing functionality. Some docstrings would help a lot. -> **The tests follow the standard structure and the functionality being tested is evident from the test names. `MM-PoE/tests/test_cli.py` has all unit tests for ` MM-PoE/mm_poe/cli.py`. Other tests e.g. for `MM-PoE/mm_poe/methods/utils/methods.py` are available in `MM-PoE/tests/methods/utils/test_methods.py`.**

There are more downloaded tests, but I only tried running the ones mentioned in the README file. All should be checked to run correctly when the ones above are fixed. -> **Please run only the scripts with numbers 7,8 and 9 as instructed in the README.**

It is not trivial to release software that people can test in arbitrary hardware, so an alternative to make things easier could be to include a Colab notebook, an AWS public AMI, or something similar. Then it is guaranteed that the reviewers will get the same results as the authors. -> **A Google Colab notebook [link](https://colab.research.google.com/github/souradipp76/MM-PoE/blob/main/scripts/experiments.ipynb) has been provided now. It has also been mentioned that the experiments should be run on CUDA-compatible hardware.**

Documentation about 2) Report issues or problems with the software, and 3) Seek support is missing. -> **Included under [Asking Questions and Reporting Issues](https://github.com/souradipp76/MM-PoE/blob/main/CONTRIBUTING.md#asking-questions-and-reporting-issues) section in the CONTRIBUTING.md

paper.md contains all the required sections and some references, but the paper itself is still not compiled and paper.bib contains many more reference entries than the ones in the draft. Will be able to assess once the compiled PDF is added. -> **To cut down on the words in the paper draft, some content was removed which removed the references from the paper but not from the bib. Also, I believe the paper compilation is done automatically by the editorialbot (see [here](https://github.com/openjournals/joss-reviews/issues/7391#issuecomment-2429142994)). If there are additional steps that are required for compiling the paper, kindly let me know.**
