### 2025-02-16 #7783 @editorialbot
[REVIEW]: Multiple Choice Reasoning via. Process of Elimination using Multi-Modal Models
**Submitting author:** @hellokayas (<a href="http://orcid.org/0009-0004-6179-389X">Sayak Chakrabarty</a>)
**Repository:** https://github.com/souradipp76/MM-PoE
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.2
