### 2025-02-16 #7783 @faroit
@jorgedch we have now started the official review phase. Please start your review check list here and checkmark what you have done already (I guess accidentally) in the pre-review issue thread.

Note that, I would highly appreciate if you would create issues in the submitted software repo and link back to this issue to reduce the amount of clutter that can happen with conversations about issues with the author. 🙏
