### 2025-02-24 souradipp76/MM-PoE#11 @hoechenberger
Installation succeeds, but running fails
Hello, I followed the installation instructions by manually creating a new vent and installing the package from PyPI via:

```shell
pip install mm_poe
```

The installation finished successfully, but running MM-PoE fails:

```python
❯ python -m mm_poe
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/mm_poe/__main__.py", line 3, in <module>
    from mm_poe.cli import main
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/mm_poe/cli.py", line 13, in <module>
    from mm_poe.methods.utils.data import (
    ...<5 lines>...
    )
ModuleNotFoundError: No module named 'mm_poe.methods'
```

This was part of my JOSS review process at https://github.com/openjournals/joss-reviews/issues/7783
