### 2025-02-24 #7783 @hoechenberger
@hellokayas  There are a few references that link to arXiv versions of publications. At least for Ma, C., & Du, X. (2023), a final version of the paper has meanwhile been published, making the arXiv preprint obsolete. Could you please check the references again and ensure you're always referencing and linking to the "properly" published pieces of work?

Thank you!
