### 2025-02-24 #7783 @hoechenberger
@hellokayas
Regarding this point:
> *  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

I'm not sure this is being sufficiently addressed right now – please correct me if I'm mistaken though! I do see that you compare **methods / algorithms** used elsewhere, but there doesn't seem to be any mention of alternative or related **software packages.** Could you please include such references, if and where applicable?
