+++ paper/paper.bib
@@ -51,7 +51,8 @@
 @article{ma2023poe,
   title={POE: Process of Elimination for Multiple Choice Reasoning},
   author={Ma, Chenkai and Du, Xinya},
-  journal={arXiv preprint arXiv:2310.15575},
+  booktitle={Proceedings of the 2023 Conference on Empirical Methods in Natural Language Processing},
+  pages={4487--4496},
   year={2023},
   doi = "10.18653/v1/2023.emnlp-main.273"
 }
@@ -160,8 +161,8 @@
 @article{shum2023automatic,
   title = {Automatic prompt augmentation and selection with chain-of-thought from labeled data},
   author = {Shum, KaShun and Diao, Shizhe and Zhang, Tong},
-  journal = {arXiv preprint arXiv:2302.12822},
-  year = {2023},
+  booktitle={The 2023 Conference on Empirical Methods in Natural Language Processing},
+  year={2023},
   doi = "10.18653/v1/2023.findings-emnlp.811"
 }
 
@@ -228,8 +229,10 @@
 @article{yao2023tree,
   title = {Tree of thoughts: Deliberate problem solving with large language models},
   author = {Yao, Shunyu and Yu, Dian and Zhao, Jeffrey and Shafran, Izhak and Griffiths, Thomas L and Cao, Yuan and Narasimhan, Karthik},
-  journal = {arXiv preprint arXiv:2305.10601},
-  year = {2023},
+  journal={Advances in neural information processing systems},
+  volume={36},
+  pages={11809--11822},
+  year={2023},
   doi = "10.48550/arXiv.2305.10601"
 }
 
@@ -325,9 +328,10 @@
 @article{Kembhavi2016ADI,
   title={A Diagram is Worth a Dozen Images},
   author={Aniruddha Kembhavi and Michael Salvato and Eric Kolve and Minjoon Seo and Hannaneh Hajishirzi and Ali Farhadi},
-  journal={ArXiv},
+  booktitle={Computer Vision--ECCV 2016: 14th European Conference, Amsterdam, The Netherlands, October 11--14, 2016, Proceedings, Part IV 14},
+  pages={235--251},
   year={2016},
-  volume={abs/1603.07396},
+  organization={Springer},
   url={https://api.semanticscholar.org/CorpusID:2682274},
   doi = "10.1007/978-3-319-46493-0_15"
 }
