### 2025-02-26 #7783 @hellokayas
> [@hellokayas](https://github.com/hellokayas) There are a few references that link to arXiv versions of publications. At least for Ma, C., & Du, X. (2023), a final version of the paper has meanwhile been published, making the arXiv preprint obsolete. Could you please check the references again and ensure you're always referencing and linking to the "final" published pieces of work?
>
> Thank you!

@hoechenberger I believe you are referring to the book or journal part in the BibTeX that contains the name of the venue where the paper got published. I checked for the ones that had Arxiv written and have changed some of them to the final publishing venue. However, there are still a few that don't show the final venue yet on Google Scholar, so I have kept them unchanged.
