### 2025-02-26 #7783 @hellokayas
> [@hellokayas](https://github.com/hellokayas) Regarding this point:
>
> > * **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
>
> I'm not sure this is being sufficiently addressed right now – please correct me if I'm mistaken though! I do see that you compare **methods / algorithms** used elsewhere, but there doesn't seem to be any mention of alternative or related **software packages.** Could you please include such references, if and where applicable?

@hoechenberger There are no alternative or related software packages. The paper [POE: Process of Elimination for Multiple Choice Reasoning](https://aclanthology.org/2023.emnlp-main.273.pdf) has provided some algorithms and methods when the dataset is only text-based. We extend those methods to the case when the dataset has both text and images.
