### 2025-02-26 90f7973
Fixed entry point and updated version.
