### 2025-02-26 souradipp76/MM-PoE#11 @souradipp76
@hoechenberger  This issue has been fixed. Please try updating the package version by doing `pip install -U mm_poe` and then running the CLI.
