### 2025-02-27 souradipp76/MM-PoE#13 @hoechenberger
Model download not working for `pip`-installed package
**Describe the bug**
I installed the package from PyPI via `pip`. When trying to run the CLI, an exception is raised because the model downloader is not shipped with the package.

**To Reproduce**
```shell
❯ uv venv --python=3.13
❯ uv pip install mm-poe
❯ uv run python -m mm_poe
? Select model family? GIT
? Select model checkpoint? microsoft/git-base-vqav2
? Select model precision? FP16
? Model output directory? ./models/
? Select scoring method? language_modeling
? Select mask strategy? below_average
? Question: Which of these countries is not located in Europe?
? Choices [comma seprated]: Germany, France, Poland, Canada
? Image Path? ./images/image.png
? Ground Truth Option: 3
```

**Expected behavior**
The model gets downloaded and is being run.

**Actual behavior**
```python
02/27/2025 08:44:33 - INFO - mm_poe.cli - Set random seed to 0.
02/27/2025 08:44:33 - INFO - mm_poe.cli - Download GIT model: microsoft/git-base-vqav2.
python: can't open file '/private/tmp/review/.venv/lib/python3.13/site-packages/mm_poe/models/model_downloaders/model_downloaders.py': [Errno 2] No such file or directory
02/27/2025 08:44:33 - INFO - mm_poe.cli - Load GIT model: microsoft/git-base-vqav2.
Traceback (most recent call last):
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/transformers/utils/hub.py", line 342, in cached_file
    resolved_file = hf_hub_download(
        path_or_repo_id,
    ...<10 lines>...
        local_files_only=local_files_only,
    )
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/huggingface_hub/utils/_validators.py", line 106, in _inner_fn
    validate_repo_id(arg_value)
    ~~~~~~~~~~~~~~~~^^^^^^^^^^^
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/huggingface_hub/utils/_validators.py", line 154, in validate_repo_id
    raise HFValidationError(
    ...<2 lines>...
    )
huggingface_hub.errors.HFValidationError: Repo id must be in the form 'repo_name' or 'namespace/repo_name': './models/GIT/microsoft/git-base-vqav2'. Use `repo_type` argument if needed.

Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/mm_poe/__main__.py", line 11, in <module>
    entry_point()
    ~~~~~~~~~~~^^
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/mm_poe/__main__.py", line 7, in entry_point
    main()
    ~~~~^^
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/mm_poe/cli.py", line 167, in main
    model, tokenizer = load_model(device, model_path, args)
                       ~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/mm_poe/methods/utils/utils.py", line 609, in load_model
    tokenizer = tokenizer_func.from_pretrained(model_path)
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/transformers/models/auto/processing_auto.py", line 260, in from_pretrained
    processor_config_file = get_file_from_repo(
        pretrained_model_name_or_path, PROCESSOR_NAME, **get_file_from_repo_kwargs
    )
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/transformers/utils/hub.py", line 496, in get_file_from_repo
    return cached_file(
        path_or_repo_id=path_or_repo,
    ...<11 lines>...
        _raise_exceptions_for_connection_errors=False,
    )
  File "/private/tmp/review/.venv/lib/python3.13/site-packages/transformers/utils/hub.py", line 408, in cached_file
    raise EnvironmentError(
        f"Incorrect path_or_model_id: '{path_or_repo_id}'. Please provide either the path to a local folder or the repo_id of a model on the Hub."
    ) from e
OSError: Incorrect path_or_model_id: './models/GIT/microsoft/git-base-vqav2'. Please provide either the path to a local folder or the repo_id of a model on the Hub.
```

The required Python module, `mm_poe/models/model_downloaders/model_downloaders.py`, is not shipping with the package:
```shell
❯ ls .venv/lib/python3.13/site-packages/mm_poe/
__init__.py __main__.py __pycache__ cli.py      methods
```

**Additional context**
This was part of my JOSS review process at https://github.com/openjournals/joss-reviews/issues/7783
