### 2025-02-28 6e88afc
Fixed CLI issue for model download.
