+++ mm_poe/methods/language_modeling.py
@@ -6,7 +6,7 @@ import os
 import torch
 from torch.utils.data import DataLoader
 
-from utils.data import (
+from mm_poe.methods.utils.data import (
     upload_to_huggingface_hub,
     preprocess_function_seq2seq,
     preprocess_function_causal,
@@ -15,7 +15,7 @@ from utils.data import (
     create_synonym_dataset,
     create_n_shot_splits,
 )
-from utils.methods import (
+from mm_poe.methods.utils.methods import (
     compute_conditional_score_seq2seq,
     compute_conditional_score_causal,
     inference_language_modeling,
@@ -23,7 +23,7 @@ from utils.methods import (
     inference_generate_synonyms,
     generate_synonyms,
 )
-from utils.utils import (
+from mm_poe.methods.utils.utils import (
     load_data,
     load_model,
     parse_args,
