+++ mm_poe/methods/vision_language_modeling.py
@@ -6,7 +6,7 @@ import os
 import torch
 from torch.utils.data import DataLoader
 
-from utils.data import (
+from mm_poe.methods.utils.data import (
     upload_to_huggingface_hub,
     preprocess_function_seq2seq,
     preprocess_function_causal,
@@ -19,7 +19,7 @@ from utils.data import (
     create_synonym_dataset,
     create_n_shot_splits,
 )
-from utils.methods import (
+from mm_poe.methods.utils.methods import (
     compute_conditional_score_seq2seq,
     compute_conditional_score_causal,
     compute_conditional_score_seq2seq_vqa,
@@ -29,7 +29,7 @@ from utils.methods import (
     inference_generate_synonyms,
     generate_synonyms,
 )
-from utils.utils import (
+from mm_poe.methods.utils.utils import (
     load_data,
     load_model,
     parse_args,
