### 2025-03-02 souradipp76/MM-PoE#14 @hoechenberger
Users should be able to download and run models without having to clone the repository
Following [the documentation](https://github.com/souradipp76/MM-PoE/blob/3b40b62d90ad1bf4a2e4f9d29875bb4f948da5aa/README.md#running-experiments) for running an experiment, the instructions start with running `make install`. However, this naturally only works for the "Install it from source" and not for the "Install it from PyPI" installation method.

FWIW, a user should never have to clone and install the latest version of a package from GitHub; installation from PyPI should be sufficient.

I'm wondering if you could convert the `.sh` scripts into Python scripts that use e.g. `argparse` to handle command line arguments, and deploy them as entry-points? This would make things much easier for users, and shouldn't be too difficult for you do implement.

This was part of my JOSS review process at https://github.com/openjournals/joss-reviews/issues/7783
