### 2025-03-02 souradipp76/MM-PoE#15 @hoechenberger
Import fails in Colab notebook
**Describe the bug**
Running the models via the provided Colab notebook fails, as `mm_poe` cannot be imported.

**To Reproduce**
* [open notebook](https://colab.research.google.com/github/souradipp76/MM-PoE/blob/main/scripts/experiments.ipynb)
* update Huggingface token
* run notebook

**Expected behavior**
Should run without error

**Actual behavior**
In the "Running experiments" section, I get plenty errors of this kind:

```python
Traceback (most recent call last):
  File "/content/MM-PoE/mm_poe/methods/vision_language_modeling.py", line 9, in <module>
    from mm_poe.methods.utils.data import (
ModuleNotFoundError: No module named 'mm_poe'
```

This was part of my JOSS review process at https://github.com/openjournals/joss-reviews/issues/7783
