### 2025-03-03 souradipp76/MM-PoE#16 @hoechenberger
Shell command that generates `USING_POETRY` raises an error message
https://github.com/souradipp76/MM-PoE/blob/781b6326c7bbc3996f1e1e6637ea77cb058513b0/Makefile#L3

yields (as expected):
```
grep: pyproject.toml: No such file or directory
```

Since you're not using poetry, I'd suggest removing all references to poetry from your code and tooling.

This was part of my JOSS review process at https://github.com/openjournals/joss-reviews/issues/7783
