### 2025-03-03 souradipp76/MM-PoE#14 @souradipp76
@hoechenberger The `.sh` scripts provided ultimately call a Python script behind the scenes which uses `argparse` to handle the command line arguments(e.g. `7_main_exp_vqa.sh` calls `vision_language_modeling.py`). These `.sh` scripts are used mainly to easily validate the package functionality and reproduce the results in the paper using some parameters set inside the script.

Also, you can use the "Install it from PyPI" installation and run the experiments but you will need the associated Python scripts for that. That's the reason why the instruction provided specifically asks users to clone the repository and run the script. Otherwise, the user have to download the Python script (say `vision_language_modeling.py`) and execute that script file with appropriate arguments after installing the package from PyPI.
