### 2025-03-03 souradipp76/MM-PoE#14 @hoechenberger
Hi @souradipp76,

thanks for your swift response!

> The `.sh` scripts provided ultimately call a Python script behind the scenes which uses `argparse` to handle the command line arguments

Yes, I realized that immediately after posting my comment here, but then didn't have time to edit my comment. Sorry about that!

> These `.sh` scripts are used mainly to easily validate the package functionality and reproduce the results in the paper using some parameters set inside the script.

Ah, so you would say they're not part of the "core" functionality of the package? In that case, it probably doesn't matter as far as the JOSS review is concerned.

Thoughts?
