### 2025-03-04 #7783 @hoechenberger
@hellokayas
For the references (again, sorry!), in the proof generated above, the references in lines 207 and 213 don't have the name of the journal / conference proceedings stated. Otherwise all is looking good to me.
