+++ paper/paper.bib
@@ -48,13 +48,22 @@
   doi = "10.1006/inco.2001.2922"
 }
 
-@article{ma2023poe,
-  title={POE: Process of Elimination for Multiple Choice Reasoning},
-  author={Ma, Chenkai and Du, Xinya},
-  booktitle={Proceedings of the 2023 Conference on Empirical Methods in Natural Language Processing},
-  pages={4487--4496},
-  year={2023},
-  doi = "10.18653/v1/2023.emnlp-main.273"
+@inproceedings{ma-du-2023-poe,
+    title = "{POE}: Process of Elimination for Multiple Choice Reasoning",
+    author = "Ma, Chenkai  and
+      Du, Xinya",
+    editor = "Bouamor, Houda  and
+      Pino, Juan  and
+      Bali, Kalika",
+    booktitle = "Proceedings of the 2023 Conference on Empirical Methods in Natural Language Processing",
+    month = dec,
+    year = "2023",
+    address = "Singapore",
+    publisher = "Association for Computational Linguistics",
+    url = "https://aclanthology.org/2023.emnlp-main.273/",
+    doi = "10.18653/v1/2023.emnlp-main.273",
+    pages = "4487--4496",
+    abstract = "Language models (LMs) are capable of conducting in-context learning for multiple choice reasoning tasks, but the options in these tasks are treated equally. As humans often first eliminate wrong options before picking the final correct answer, we argue a similar two-step strategy can make LMs better at these tasks. To this end, we present the Process of Elimination (POE), a two-step scoring method. In the first step, POE scores each option, and eliminates seemingly wrong options. In the second step, POE masks these wrong options, and makes the final prediction from the remaining options. Zero-shot experiments on 8 reasoning tasks illustrate the effectiveness of POE, and a following analysis finds our method to be especially performant on logical reasoning tasks. We further analyze the effect of masks, and show that POE applies to few-shot settings and large language models (LLMs) like ChatGPT."
 }
 
 @inproceedings{holtzman2021surface,
@@ -325,7 +334,7 @@
   doi = "10.1109/ICCV.2015.279"
 }
 
-@article{Kembhavi2016ADI,
+@inproceedings{Kembhavi2016ADI,
   title={A Diagram is Worth a Dozen Images},
   author={Aniruddha Kembhavi and Michael Salvato and Eric Kolve and Minjoon Seo and Hannaneh Hajishirzi and Ali Farhadi},
   booktitle={Computer Vision--ECCV 2016: 14th European Conference, Amsterdam, The Netherlands, October 11--14, 2016, Proceedings, Part IV 14},
