### 2025-03-05 #7783 @hellokayas
> [@hellokayas](https://github.com/hellokayas) For the references (again, sorry!), in the proof generated above, the references in lines 207 and 213 don't have the name of the journal / conference proceedings stated. Otherwise all is looking good to me.

@hoechenberger I have generated the pdf above and you can see that the issue is fixed. So does this complete your entire checklist?
