### 2025-03-09 #7783 @jorgedch
I'll update the previous feedback based on these replies:

https://github.com/openjournals/joss-reviews/issues/7391#issuecomment-2615005522

https://github.com/openjournals/joss-reviews/issues/7391#issuecomment-2629560468
