### 2025-03-09 #7783 @faroit
@jorgedch its good that you copied in the your efforts you already did in the pre-review, but please don't edit the review checklist so that its easier for me to follow the status of the review.

- can you please revert the changes to https://github.com/openjournals/joss-reviews/issues/7783#issuecomment-2709033154
- if suitable, add your pre-review comments to separate issues in https://github.com/souradipp76/MM-PoE and in there link back to this review thread
- put general comments here

thanks a lot!
