### 2025-03-12 #7783 @hellokayas
@faroit Here are the details.

- Double check authors and affiliations (including ORCIDs) -- Done
- Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.  -- Version 1.0.4
- Archive the release on Zenodo/figshare/etc and post the DOI here.  -- DOI  =  https://doi.org/10.5281/zenodo.15009921
- Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.  -- Done
- Make sure that the license listed for the archive is the same as the software license.  -- Apache 2.0
