### 2025-03-19 51e2bd3
Update missing reference in paper.md
