+++ paper/paper.md
@@ -43,7 +43,7 @@ This paper argues that vision language models can benefit from an explicit two-s
 
 Our hypothesis posits that vision language models, when equipped with a mechanism to discard implausible answers systematically, can achieve better performance on multiple-choice visual reasoning tasks. This is particularly relevant in the context of logical reasoning, where the elimination of clearly incorrect options can simplify the decision process and potentially lead to more accurate outcomes. This idea is supported by previous work demonstrating the effectiveness of LMs in various reasoning tasks when adapted to more human-like reasoning methods [@holtzman2021surface].
 
-In the development of MM-PoE, we draw inspiration from the established capabilities of LMs to handle complex reasoning tasks [@brown2020language] and the known strategies that humans employ in test-taking scenarios as depicted in [@MISSINGREF]. The approach builds on the foundational work in language modeling likelihood [@brown2020language], which demonstrates the LMs' ability to perform in-context learning. By incorporating a structured process to eliminate unlikely choices in a multi-modal setting, MM-PoE aims to refine this capability, making it more targeted and efficient in dealing with the nuanced challenges presented by multiple-choice questions.
+In the development of MM-PoE, we draw inspiration from the established capabilities of LMs to handle complex reasoning tasks [@brown2020language] and the known strategies that humans employ in test-taking scenarios as depicted in [@ma-du-2023-poe]. The approach builds on the foundational work in language modeling likelihood [@brown2020language], which demonstrates the LMs' ability to perform in-context learning. By incorporating a structured process to eliminate unlikely choices in a multi-modal setting, MM-PoE aims to refine this capability, making it more targeted and efficient in dealing with the nuanced challenges presented by multiple-choice questions.
 
 The effectiveness of this approach is underscored through zero-shot and few-shot experiments across a diverse set of reasoning datasets, illustrating that the integration of human-like elimination strategies can significantly enhance the performance of vision language models. This paper aims to show that by mimicking human reasoning processes, we can make VLMs not only perform better on standardized visual reasoning tasks but also behave in ways that are more interpretable and aligned with human cognitive processes.
 
