### 2025-03-19 96fae46
Added doi for Idefics2 and updated paper.bib
