+++ paper/paper.bib
@@ -263,70 +263,71 @@
   doi = "10.48550/arXiv.2102.09690"
 }
 
-@software{BLIP2_Opt,
+@misc{BLIP2_Opt,
   author = {{Salesforce}},
   title = {BLIP2: Opt Model},
   version = {2.7b},
   howpublished = {\url{https://huggingface.co/Salesforce/blip2-opt-2.7b}}
 }
 
-@conj{BLIP2_FLAN,
+@misc{BLIP2_FLAN,
   author = {{Salesforce}},
   title = {BLIP2: FLAN-T5-XL Model},
   version = {XL},
   howpublished = {\url{https://huggingface.co/Salesforce/blip2-flan-t5-xl}}
 }
 
-@conj{InstructBLIP,
+@misc{InstructBLIP,
   author = {{Salesforce}},
   title = {InstructBLIP: Vicuna Model},
   version = {7b},
   howpublished = {\url{https://huggingface.co/Salesforce/instructblip-vicuna-7b}}
 }
 
-@conj{GIT_Base_VQA,
+@misc{GIT_Base_VQA,
   author = {{Microsoft}},
   title = {GIT: Base VQA Model},
   version = {Base},
   howpublished = {\url{https://huggingface.co/microsoft/git-base-vqav2}}
 }
 
-@conj{GIT_Base_TextVQA,
+@misc{GIT_Base_TextVQA,
   author = {{Microsoft}},
   title = {GIT: Base TextVQA Model},
   version = {Base},
   howpublished = {\url{https://huggingface.co/microsoft/git-base-textvqa}}
 }
 
-@conj{PaliGemma_ScienceQA,
+@misc{PaliGemma_ScienceQA,
   author = {{Google}},
   title = {PaliGemma: Fine-Tuned Science QA Model},
   version = {3b},
   howpublished = {\url{https://huggingface.co/google/paligemma-3b-ft-science-qa-448}}
 }
 
-@conj{PaliGemma_VQAV2,
+@misc{PaliGemma_VQAV2,
   author = {{Google}},
   title = {PaliGemma: Fine-Tuned VQAV2 Model},
   version = {3b},
   howpublished = {\url{https://huggingface.co/google/paligemma-3b-ft-vqav2-448}}
 }
 
-@conj{PaliGemma_AI2D,
+@misc{PaliGemma_AI2D,
   author = {{Google}},
   title = {PaliGemma: Fine-Tuned AI2D Model},
   version = {3b},
   howpublished = {\url{https://huggingface.co/google/paligemma-3b-ft-ai2d-448}}
 }
 
-@conj{Idefics2,
+@misc{Idefics2,
   author = {{Hugging Face}},
   title = {Idefics2: 8b Model},
   version = {8b},
-  howpublished = {\url{https://huggingface.co/HuggingFaceM4/idefics2-8b}}
+  howpublished = {\url{https://huggingface.co/HuggingFaceM4/idefics2-8b}},
+  doi = "10.48550/arXiv.2405.02246"
 }
 
-@InProceedings{VQA,
+@inproceedings{VQA,
   author = {Stanislaw Antol and Aishwarya Agrawal and Jiasen Lu and Margaret Mitchell and Dhruv Batra and C. Lawrence Zitnick and Devi Parikh},
   title = {VQA: Visual Question Answering},
   booktitle = {International Conference on Computer Vision (ICCV)},
