### 2025-03-19 #7783 @hellokayas
> [@hellokayas](https://github.com/hellokayas) this looks great now! We are all good

@faroit Thank you so much! Will we find our paper on the JOSS published papers page(https://joss.theoj.org/papers/published) within a few days?
