### 2025-03-22 #7783 @hellokayas
@faroit We think one last step remains, i.e., the editor will have to use the @ editorial bot accept command. Please do that whenever you find time. Let us know if we need to do something from our side.
