### 2025-04-03 #7783 @hellokayas
@crvernon Did you have a chance to proceed with your checklist?
