### 2025-04-07 #7783 @crvernon
@hellokayas - Please address the following before I accept this for publication:

In your archive:

- [ ] The author list order of your archive must match the order listed in your paper exactly.  Please edit the metadata in your archive to address this.  Note:  a new release is not necessary.

In your paper:

- [ ] LINE 35:  Missing space between "strategies" and "(Zhang et al., 2023)"
- [ ] LINE 71:  Missing space between "mechanism" and "(Datta & Chakrabarty, 2024)"
- [ ] LINE 115:  Missing space between "Answering" and "(VQA)"
- [ ] LINE 116:  Missing space between "Understanding" and "(AI2D)"

Let me know when these have been fixed.  Thanks.
