### 2025-04-14 #7783 @faroit
🥳 @hellokayas congrats also from my side. I am sorry that the turnaround time wasn't ideal this time which was also partly on my side as I wasn't been able to be as responsive as I use to be. I hope you still like the overall outcome and the "JOSS experience".

@jorgedch and @hoechenberger thank your for your excellent reviews. I think they help to significantly improve the paper
